modify signed to unsigned

+++ services/distributedfiledaemon/src/network/softbus/softbus_handler.cpp
@@ -256,7 +256,7 @@ void SoftBusHandler::SetSocketOpt(int32_t socketId, const char **src, uint32_t s
         if (stat(file, &buf) == -1) {
             return;
         }
-        totalSize += buf.st_size;
+        totalSize += static_cast<uint64_t>(buf.st_size);
     }
 
     TransFlowInfo flowInfo;
